fixup: tweaks and docs improvements
