+++ README.md
@@ -26,9 +26,6 @@ The Decoy library allows you to create, stub, and verify fully-typed, async/awai
 
 The Decoy API is heavily inspired by / stolen from the excellent [testdouble.js][] and [Mockito][] projects.
 
-[testdouble.js]: https://github.com/testdouble/testdouble.js
-[mockito]: https://site.mockito.org/
-
 ## Install
 
 ```bash
@@ -55,18 +52,92 @@ def test_my_thing_works(decoy: Decoy) -> None:
 
 The `decoy` fixture is function-scoped and will ensure that all stub and spy state is reset between every test.
 
-[pytest]: https://docs.pytest.org/
-
 ### Mypy Setup
 
 Decoy's API can be a bit confusing to [mypy][]. To suppress mypy errors that may be emitted during valid usage of the Decoy API, we have a mypy plugin that you should add to your configuration file:
 
 ```ini
-# mypi.ini
+# mypy.ini
 
 # ...
 plugins = decoy.mypy
 # ...
 ```
 
+## Basic Usage
+
+This example assumes you are using [pytest][]. See Decoy's [documentation][] for a more detailed usage guide and API reference.
+
+### Define your test
+
+Decoy will add a `decoy` fixture that provides its mock creation API.
+
+```python
+from decoy import Decoy
+from todo import TodoAPI, TodoItem
+from todo.store TodoStore
+
+def test_add_todo(decoy: Decoy) -> None:
+    ...
+```
+
+### Create a mock
+
+Use `decoy.create_decoy` to create a mock based on some specification. From there, inject the mock into your test subject.
+
+```python
+def test_add_todo(decoy: Decoy) -> None:
+    todo_store = decoy.create_decoy(spec=TodoStore)
+    subject = TodoAPI(store=todo_store)
+    ...
+```
+
+See [creating mocks][] for more details.
+
+### Stub a behavior
+
+Use `decoy.when` to configure your mock's behaviors. For example, you can set the mock to return a certain value when called in a certain way using `then_return`:
+
+```python
+def test_add_todo(decoy: Decoy) -> None:
+    """Adding a todo should create a TodoItem in the TodoStore."""
+    todo_store = decoy.create_decoy(spec=TodoStore)
+    subject = TodoAPI(store=todo_store)
+
+    decoy.when(
+        todo_store.add(name="Write a test for adding a todo")
+    ).then_return(
+        TodoItem(id="abc123", name="Write a test for adding a todo")
+    )
+
+    result = subject.add("Write a test for adding a todo")
+    assert result == TodoItem(id="abc123", name="Write a test for adding a todo")
+```
+
+See [stubbing with when][] for more details.
+
+### Verify a call
+
+Use `decoy.verify` to assert that a mock was called in a certain way. This is best used with dependencies that are being used for their side-effects and don't return a useful value.
+
+```python
+def test_remove_todo(decoy: Decoy) -> None:
+    """Removing a todo should remove the item from the TodoStore."""
+    todo_store = decoy.create_decoy(spec=TodoStore)
+    subject = TodoAPI(store=todo_store)
+
+    subject.remove("abc123")
+
+    decoy.verify(todo_store.remove(id="abc123"))
+```
+
+See [spying with verify][] for more details.
+
+[testdouble.js]: https://github.com/testdouble/testdouble.js
+[mockito]: https://site.mockito.org/
+[pytest]: https://docs.pytest.org/
 [mypy]: https://mypy.readthedocs.io/
+[documentation]: https://mike.cousins.io/decoy/
+[creating mocks]: https://mike.cousins.io/decoy/usage/create/
+[stubbing with when]: https://mike.cousins.io/decoy/usage/when/
+[spying with verify]: https://mike.cousins.io/decoy/usage/verify/
