+++ docs/usage/create.md
@@ -7,9 +7,9 @@ Decoy can create two kinds of mocks:
 
 ## Mocking a class
 
-The [`create_decoy` method][decoy.Decoy.create_decoy] is used to create mock class instances. Decoy will inspect type annotations and method signatures to automatically configure methods as synchronous or asynchronous. Decoy mocks are automatically deep.
+To create a mock class instance, use [decoy.Decoy.create_decoy][]. Decoy will inspect type annotations and method signatures to automatically configure methods as synchronous or asynchronous. Decoy mocks are automatically deep.
 
-To typecheckers, the mock will appear to have the exact same type as the `spec` argument. The mock will also pass `isinstance` checks.
+To type checkers, the mock will appear to have the exact same type as the `spec` argument. The mock will also pass `isinstance` checks.
 
 ```python
 def test_my_thing(decoy: Decoy) -> None:
@@ -18,9 +18,9 @@ def test_my_thing(decoy: Decoy) -> None:
 
 ## Mocking a function
 
-The [`create_decoy_func` method][decoy.Decoy.create_decoy_func] is used to create a mock function instance. Decoy can inspect a `spec` function signature to automatically configure the function as syncronous or asynchronous. Otherwise, the `is_async` argument can force the mock to be asynchronous.
+To create a mock function, use [decoy.Decoy.create_decoy_func][]. Decoy can inspect a `spec` function signature to automatically configure the function as synchronous or asynchronous. Otherwise, the `is_async` argument can force the mock to be asynchronous.
 
-To typecheckers, the mock will appear to have the exact same type as the `spec` argument, if used.
+To type checkers, the mock will appear to have the exact same type as the `spec` argument, if used.
 
 ```python
 def test_my_thing(decoy: Decoy) -> None:
