+++ tests/test_call_handler.py
@@ -2,10 +2,10 @@
 import pytest
 
 from decoy import Decoy
+from decoy.spy import SpyCall
 from decoy.call_stack import CallStack
 from decoy.stub_store import StubStore, StubBehavior
 from decoy.call_handler import CallHandler
-from decoy.call_stack import SpyCall
 
 
 @pytest.fixture
